Fix concurrent access bug
