+++ api/src/main/java/com/example/poipricing/api/config/WorkbookConfig.java
@@ -1,0 +1,14 @@
+package com.example.poipricing.api.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class WorkbookConfig {
+	@Bean
+  public SyncWorkbook getWorkbook() throws Exception {
+		SyncWorkbook syncWorkbook = new SyncWorkbook();
+		
+    return syncWorkbook;
+  }
+}
